+++ src/main/java/org/kumoricon/site/HomePresenter.java
@@ -2,7 +2,6 @@ package org.kumoricon.site;
 
 import org.kumoricon.model.badge.Badge;
 import org.kumoricon.model.badge.BadgeRepository;
-import org.kumoricon.site.HomeView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +21,7 @@ public class HomePresenter {
     private static final Logger log = LoggerFactory.getLogger(HomePresenter.class);
 
     public void showBadges(HomeView view) {
-        log.info("{} viewed home screen", view.getCurrentUser());
+        log.info("{} viewed home screen", view.getCurrentUsername());
         List<Badge> badges = badgeRepository.findByVisibleTrue();
         List<Badge> badgesUserCanView = new ArrayList<>();
         for (Badge badge : badges) {