+++ src/main/java/org/kumoricon/site/attendee/AttendeePrintView.java
@@ -1,7 +1,6 @@
 package org.kumoricon.site.attendee;
 
 import org.kumoricon.model.attendee.Attendee;
-import org.kumoricon.model.user.User;
 
 import java.util.List;
 
@@ -9,6 +8,6 @@ public interface AttendeePrintView {
     void showPrintBadgeWindow(List<Attendee> attendeeList);
     void notify(String message);
     void notifyError(String message);
-    User getCurrentUser();
+    String getCurrentUsername();
     String getCurrentClientIPAddress();
 }