+++ src/main/java/org/kumoricon/site/attendee/prereg/PreRegPresenter.java
@@ -42,7 +42,7 @@ public class PreRegPresenter implements PrintBadgeHandler {
     }
 
     public void searchFor(String searchString) {
-        log.info("{} searched preregistered attendees for {}", view.getCurrentUser(), searchString);
+        log.info("{} searched preregistered attendees for {}", view.getCurrentUsername(), searchString);
         view.getAttendeeBeanList().removeAllItems();
         if (searchString != null && !searchString.trim().isEmpty()) {
             searchString = searchString.trim();
@@ -61,7 +61,7 @@ public class PreRegPresenter implements PrintBadgeHandler {
     }
 
     public void selectAttendee(Attendee attendee) {
-        log.info("{} viewed preregistered attendee {}", view.getCurrentUser(), attendee);
+        log.info("{} viewed preregistered attendee {}", view.getCurrentUsername(), attendee);
         if (attendee.getBadge().getWarningMessage() == null) {
             continueCheckIn(attendee);
         } else {
@@ -81,7 +81,7 @@ public class PreRegPresenter implements PrintBadgeHandler {
             view.showAttendee(attendee, badgeRepository.findAll());
         } else {
             log.error("{} tried to view preregistered attendee id {} but they were not found",
-                    view.getCurrentUser(), id);
+                    view.getCurrentUsername(), id);
         }
     }
 
@@ -134,7 +134,7 @@ public class PreRegPresenter implements PrintBadgeHandler {
 
     @Override
     public void showAttendeeBadgeWindow(AttendeePrintView view, List<Attendee> attendeeList) {
-        log.info("{} printing badges for preregistered attendee(s) {}", view.getCurrentUser(), attendeeList);
+        log.info("{} printing badges for preregistered attendee(s) {}", view.getCurrentUsername(), attendeeList);
         view.notify(badgePrintService.printBadgesForAttendees(attendeeList, view.getCurrentClientIPAddress()));
         view.showPrintBadgeWindow(attendeeList);
     }
@@ -142,7 +142,7 @@ public class PreRegPresenter implements PrintBadgeHandler {
     @Override
     public void badgePrintSuccess(PrintBadgeWindow window, List<Attendee> attendees) {
         log.info("{} reports badges printed successfully for preregistered attendee(s) {}",
-                view.getCurrentUser(), attendees);
+                view.getCurrentUsername(), attendees);
         for (Attendee attendee : attendees) {
             attendee.setCheckedIn(true);
         }